Reorganize cernbox-specific packages
