+++ cmd/revad/runtime/loader.go
@@ -29,6 +29,7 @@ import (
 	_ "github.com/cs3org/reva/internal/http/services/loader"
 	_ "github.com/cs3org/reva/pkg/auth/manager/loader"
 	_ "github.com/cs3org/reva/pkg/auth/registry/loader"
+	_ "github.com/cs3org/reva/pkg/cbox/loader"
 	_ "github.com/cs3org/reva/pkg/metrics/driver/loader"
 	_ "github.com/cs3org/reva/pkg/ocm/invite/manager/loader"
 	_ "github.com/cs3org/reva/pkg/ocm/provider/authorizer/loader"
