+++ pkg/cbox/loader/loader.go
@@ -19,8 +19,7 @@
 package loader
 
 import (
-	// Load core share manager drivers.
-	_ "github.com/cs3org/reva/pkg/share/manager/json"
-	_ "github.com/cs3org/reva/pkg/share/manager/memory"
-	// Add your own here
+	// Load cbox specific drivers.
+	_ "github.com/cs3org/reva/pkg/cbox/share/sql"
+	_ "github.com/cs3org/reva/pkg/cbox/user/rest"
 )
